feat(plots): refactor time, index, and waterfall plots
